+++ src/api/PDFImage.ts
@@ -1,7 +1,9 @@
 import type { Embeddable } from 'src/api/Embeddable.js';
 import { PDFDocument } from 'src/api/PDFDocument.js';
-import { JpegEmbedder, PDFRef, PngEmbedder } from 'src/core/index.js';
-import { assertIs } from 'src/utils/index.js';
+import { JpegEmbedder } from 'src/core/embedders/JpegEmbedder.js';
+import { PngEmbedder } from 'src/core/embedders/PngEmbedder.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
+import { assertIs } from 'src/utils/validators.js';
 
 export type ImageEmbedder = JpegEmbedder | PngEmbedder;
