+++ src/api/form/PDFDropdown.ts
@@ -14,20 +14,18 @@ import {
 import { rgb } from 'src/api/colors.js';
 import { degrees } from 'src/api/rotations.js';
 
-import {
-  PDFHexString,
-  PDFRef,
-  PDFString,
-  PDFStream,
-  PDFWidgetAnnotation,
-  PDFAcroComboBox,
-  AcroChoiceFlags,
-} from 'src/core/index.js';
+import { PDFAcroComboBox } from 'src/core/acroform/PDFAcroComboBox.js';
+import { AcroChoiceFlags } from 'src/core/acroform/flags.js';
+import { PDFWidgetAnnotation } from 'src/core/annotation/PDFWidgetAnnotation.js';
+import { PDFHexString } from 'src/core/objects/PDFHexString.js';
+import { PDFRef } from 'src/core/objects/PDFRef.js';
+import { PDFStream } from 'src/core/objects/PDFStream.js';
+import { PDFString } from 'src/core/objects/PDFString.js';
 import {
   assertIs,
   assertOrUndefined,
   assertPositive,
-} from 'src/utils/index.js';
+} from 'src/utils/validators.js';
 
 /**
  * Represents a dropdown field of a [[PDFForm]].
