+++ src/api/operators.ts
@@ -1,12 +1,10 @@
 import { asNumber, asPDFName, asPDFNumber } from 'src/api/objects.js';
 import { degreesToRadians } from 'src/api/rotations.js';
-import {
-  PDFHexString,
-  PDFName,
-  PDFNumber,
-  PDFOperator,
-  PDFOperatorNames as Ops,
-} from 'src/core/index.js';
+import { PDFHexString } from 'src/core/objects/PDFHexString.js';
+import { PDFName } from 'src/core/objects/PDFName.js';
+import { PDFNumber } from 'src/core/objects/PDFNumber.js';
+import { PDFOperator } from 'src/core/operators/PDFOperator.js';
+import { PDFOperatorNames as Ops } from 'src/core/operators/PDFOperatorNames.js';
 
 /* ==================== Clipping Path Operators ==================== */
