+++ src/api/rotations.ts
@@ -1,5 +1,5 @@
 import { InvalidRotationError } from 'src/api/errors.js';
-import { assertIs } from 'src/utils/index.js';
+import { assertIs } from 'src/utils/validators.js';
 
 export enum RotationTypes {
   Degrees = 'degrees',
