+++ src/core/embedders/CustomFontEmbedder.ts
@@ -3,7 +3,7 @@ import type { TTFFont } from '@denkiyagi/fontkit';
 
 import { AbstractCustomFontEmbedder } from 'src/core/embedders/AbstractCustomFontEmbedder.js';
 import { InvalidFontTypeError } from 'src/core/errors.js';
-import type { EmbedFontAdvancedOptions } from 'src/api/index.js';
+import type { EmbedFontAdvancedOptions } from 'src/api/PDFDocumentOptions.js';
 
 export class CustomFontEmbedder extends AbstractCustomFontEmbedder {
   static for(
