+++ src/core/embedders/StandardFontEmbedder.ts
@@ -9,7 +9,7 @@ import { PDFHexString } from 'src/core/objects/PDFHexString.js';
 import type { PDFRef } from 'src/core/objects/PDFRef.js';
 import type { PDFContext } from 'src/core/PDFContext.js';
 import type { SingleLineTextOrGlyphs } from 'src/types/text.js';
-import { toCodePoint, toHexString } from 'src/utils/index.js';
+import { toCodePoint, toHexString } from 'src/utils/strings.js';
 
 export interface Glyph {
   code: number;
