+++ src/core/parser/PDFObjectStreamParser.ts
@@ -5,7 +5,7 @@ import type { PDFRawStream } from 'src/core/objects/PDFRawStream.js';
 import { PDFRef } from 'src/core/objects/PDFRef.js';
 import { ByteStream } from 'src/core/parser/ByteStream.js';
 import { PDFObjectParser } from 'src/core/parser/PDFObjectParser.js';
-import { waitForTick } from 'src/utils/index.js';
+import { waitForTick } from 'src/utils/async.js';
 
 export class PDFObjectStreamParser extends PDFObjectParser {
   static forStream = (
