+++ src/core/structures/PDFCatalog.ts
@@ -3,7 +3,7 @@ import { PDFName } from 'src/core/objects/PDFName.js';
 import type { PDFRef } from 'src/core/objects/PDFRef.js';
 import type { PDFContext } from 'src/core/PDFContext.js';
 import type { PDFPageTree } from 'src/core/structures/PDFPageTree.js';
-import { PDFAcroForm } from 'src/core/acroform/index.js';
+import { PDFAcroForm } from 'src/core/acroform/PDFAcroForm.js';
 import { ViewerPreferences } from '../interactive/ViewerPreferences.js';
 
 export class PDFCatalog extends PDFDict {
